fix: final audit — all 40 remaining findings resolved
Performance:
- IdentityHasher for stmt cache (pre-hashed keys, no SipHash overhead)
- Column offsets (usize,i32) → (u32,i32): halved metadata memory per row
- ErrorFields severity/code: Box<str> instead of String
- pg_code: Option<Box<str>> instead of Option<String>
- DecodeError: Cow<'static, str> for column/actual
- TLS ClientConfig cached in LazyLock<Arc> (built once, not per connection)
- Levenshtein: fixed [usize; 128] stack array (no heap alloc)
- Server params: Vec<(Box<str>, Box<str>)> replacing HashMap
- MD5 password: [u8; 36] fixed array, hex_encode: [u8; 32]
- SASL mechanisms: SmallVec<[&str; 2]>
- Decimal encode: SmallVec intermediates
- Decimal decode: arithmetic instead of string format+parse
- normalize_sql: last_was_space flag instead of ends_with(' ')
- Notification buffer: 10000 → 1024
- write_buf shrink threshold: 4KB → 8KB
- Cached gen_row_decode in codegen (no duplicate TokenStream)
- Extracted gen_stream_struct (shared across 3 codegen paths)
- param_oids: SmallVec<[u32; 8]>
- make_stmt_name helper (was format! repeated 4 times)
- seen_names: HashSet instead of Vec::contains (O(1) vs O(n))

Security:
- Password zeroized after PBKDF2 derivation
- #[inline(never)] on constant_time_eq
- resolve_offset i==0 underflow guard

Correctness:
- chrono DateTime path match includes ::chrono::Utc variant
- Unified ColumnInfo/ColumnDesc (deleted duplicate proto struct)
- Removed double config.validate() call
- warmup_sqls: Mutex instead of RwLock (simpler, rare access)

DX:
- Pool Clone impl in bsql-core
- CI: RUSTFLAGS="-D warnings" replacing fragile grep pipeline

Test:
- Arena resolve_offset with global_offset=0
- url_decode multi-byte UTF-8 (%C3%A9 → é)

593 tests, all passing.

+++ crates/bsql-driver/tests/integration.rs
@@ -1096,7 +1096,7 @@ async fn streaming_1000_rows() {
 
     loop {
         let num_cols = columns.len();
-        let mut col_offsets: Vec<(usize, i32)> = Vec::new();
+        let mut col_offsets: Vec<(u32, i32)> = Vec::new();
 
         if !first_chunk {
             conn.streaming_send_execute(64).await.unwrap();
@@ -1117,7 +1117,7 @@ async fn streaming_1000_rows() {
         for i in 0..row_count {
             let (offset, len) = col_offsets[i * num_cols];
             if len >= 0 {
-                let data = arena.get(offset, len as usize);
+                let data = arena.get(offset as usize, len as usize);
                 let val = i32::from_be_bytes([data[0], data[1], data[2], data[3]]);
                 all_values.push(val);
             }
@@ -1157,7 +1157,7 @@ async fn streaming_chunk_boundary_exact() {
         .unwrap();
 
     let num_cols = columns.len();
-    let mut col_offsets: Vec<(usize, i32)> = Vec::new();
+    let mut col_offsets: Vec<(u32, i32)> = Vec::new();
     let more = conn
         .streaming_next_chunk(&mut arena, &mut col_offsets)
         .await
@@ -1203,7 +1203,7 @@ async fn streaming_zero_rows() {
         .unwrap();
 
     let num_cols = columns.len();
-    let mut col_offsets: Vec<(usize, i32)> = Vec::new();
+    let mut col_offsets: Vec<(u32, i32)> = Vec::new();
     let more = conn
         .streaming_next_chunk(&mut arena, &mut col_offsets)
         .await
@@ -1234,7 +1234,7 @@ async fn streaming_single_row() {
         .unwrap();
 
     let num_cols = columns.len();
-    let mut col_offsets: Vec<(usize, i32)> = Vec::new();
+    let mut col_offsets: Vec<(u32, i32)> = Vec::new();
     let more = conn
         .streaming_next_chunk(&mut arena, &mut col_offsets)
         .await
@@ -1246,7 +1246,7 @@ async fn streaming_single_row() {
 
     let (offset, len) = col_offsets[0];
     assert_eq!(len, 4);
-    let data = arena.get(offset, len as usize);
+    let data = arena.get(offset as usize, len as usize);
     let val = i32::from_be_bytes([data[0], data[1], data[2], data[3]]);
     assert_eq!(val, 42);
 }
@@ -1268,7 +1268,7 @@ async fn streaming_early_drop() {
         .await
         .unwrap();
 
-    let mut col_offsets: Vec<(usize, i32)> = Vec::new();
+    let mut col_offsets: Vec<(u32, i32)> = Vec::new();
     let more = guard
         .streaming_next_chunk(&mut arena, &mut col_offsets)
         .await
